+++ Hassio-pacman-cherry/Scripts/builder.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/with-contenv bashio
+#!/usr/bin/env bash
 
 # Cargar variables de entorno globales
 source /env.sh