+++ Hassio-pacman-cherry/Scripts/run.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/with-contenv bashio
+#!/usr/bin/env bash
 set -euo pipefail
 
 # Cargar variables de entorno globales