+++ Hassio-pacman-cherry/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## 0.22
+
+- changes related with vaiables
 
 ## 0.2.1